backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.123.5",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.2.1",
    "langchain>=1.1.2",
    "langgraph>=1.0.4",
    "langchain-google-genai>=1.0.0",
    "tavily-python>=0.7.14",
    "pydantic>=2.12.5",
    "langchain-openai>=1.1.1",
    "langchain-deepseek>=1.0.1",
]
